update mac libs to universal versions
Includes:

OpenAL Soft 1.23.1
The system OpenAL framework from Apple is deprecated and is missing several
features, such as working EFX. This is a modified openal-soft framework that
mimics the Apple version. The modifications are packaging related only, changing
the framework lib name from "soft_oal" to "OpenAL" and the version from "C"
to "A".

ffmpeg 3.4.13
Newest stable release from the 3.4 version. Old, but basically matches the
releases currently in use by FSO.

freetype 2.10.4
Same version currently in use but recompiled to be a universal lib. To get it
linking properly on both architectures required switching from a static lib
to a dynamic one. Linked against the static libs for png and zlib included with
FSO, and with bzip2, brotli, and harfbuzz disabled.